[Fixture #160806671] Rename folders to v2
