--- app/api/v1/__init__.py
@@ -1,1 +1,0 @@
-"""Make v1 a module"""
